python/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

# PyTorch
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121

# WEB server dependencies
Flask==3.1.0
flask-cors==5.0.0

# FAISS
tf-keras==2.19.0
faiss-cpu==1.11.0
numpy<2
scipy==1.13.1

# Audio, imagen and video
Pillow==10.4.0
librosa==0.10.0
openai-whisper==20240930
decord==0.6.0
moviepy==2.2.1

# MongoDB
pymongo<4.0

# Documents: PDF, Word, Excel, PowerPoint
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.13
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1
python-docx==1.1.2
python-pptx==1.0.2

# ImageBind Dependencies
transformers==4.41.2
pytorch-lightning==2.2.5
opencv-python==4.10.0.82 
tqdm==4.66.4
scikit-learn==1.5.0
ftfy==6.1.1
regex==2023.12.25
termcolor==2.4.0
